:construction: move project display to client side
