+++ app/projects/[slug]/page.tsx
@@ -1,9 +1,6 @@
 import { createServerClient } from '@/db/supabase-server'
 import { getUser, isAdmin, getProfileAndBidsById } from '@/db/profile'
-import { PlaceBid } from './place-bid'
-import { RichContent } from '@/components/editor'
 import { CloseBidding } from './close-bidding'
-import { EditDescription } from './edit-description'
 import { getFullProjectBySlug, getProjectBySlug } from '@/db/project'
 import { getCommentsByProject } from '@/db/comment'
 import { getBidsByProject } from '@/db/bid'
@@ -13,23 +10,15 @@ import {
   getAmountRaised,
   getProposalValuation,
 } from '@/utils/math'
-import { ProposalData } from './proposal-data'
-import { SignInButton } from '@/components/sign-in-button'
-import { ProjectTabs } from './project-tabs'
 import { getTxnsByProject, getTxnsByUser } from '@/db/txn'
-import { Description } from './description'
-import { ProjectCardHeader } from '@/components/project-card'
 import {
   calculateSellableShares,
   calculateCharityBalance,
   calculateCashBalance,
 } from '@/utils/math'
-import { DonateBox } from '@/components/donate-box'
-import { Divider } from '@/components/divider'
-import { ProposalRequirements } from './proposal-requirements'
-import { ProgressBar } from '@/components/progress-bar'
 import { getUserEmail } from '@/utils/email'
 import { createAdminClient } from '@/pages/api/_db'
+import { ProjectDisplay } from './project-display'
 
 export const revalidate = 0
 
@@ -52,7 +41,7 @@ export default async function ProjectPage(props: { params: { slug: string } }) {
   if (!project) {
     return <div>404: Project not found.</div>
   }
-  const [profile, userTxns, comments, projectBids, projectTxns] =
+  const [userProfile, userTxns, comments, projectBids, projectTxns] =
     await Promise.all([
       user ? await getProfileAndBidsById(supabase, user.id) : null,
       user ? getTxnsByUser(supabase, user.id) : [],
@@ -60,120 +49,21 @@ export default async function ProjectPage(props: { params: { slug: string } }) {
       getBidsByProject(supabase, project.id),
       getTxnsByProject(supabase, project.id),
     ])
-  const creatorEmail = profile?.regranter_status
+  const creatorEmail = userProfile?.regranter_status
     ? await getUserEmail(createAdminClient(), project.creator)
     : undefined
-  const userSpendableFunds = profile
-    ? profile.accreditation_status && project.type === 'cert'
-      ? calculateCashBalance(userTxns, profile.bids, profile.id, true)
-      : calculateCharityBalance(
-          userTxns,
-          profile.bids,
-          profile.id,
-          profile.accreditation_status
-        )
-    : 0
-  const userSellableShares = profile
-    ? calculateSellableShares(userTxns, profile.bids, project.id, profile.id)
-    : 0
-  const userShares = profile
-    ? calculateShares(userTxns, project.id, profile.id)
-    : 0
-  const valuation =
-    project.type === 'grant'
-      ? project.funding_goal
-      : project.stage === 'proposal'
-      ? getProposalValuation(project)
-      : getActiveValuation(
-          projectTxns,
-          projectBids,
-          getProposalValuation(project)
-        )
-  const isOwnProject = user?.id === project.profiles.id
-  const pendingProjectTransfers = project.project_transfers?.filter(
-    (projectTransfer) => !projectTransfer.transferred
-  )
-  const amountRaised = getAmountRaised(project, projectBids, projectTxns)
   return (
     <>
-      {project.type === 'grant' &&
-        project.stage === 'proposal' &&
-        pendingProjectTransfers.length === 0 && (
-          <ProposalRequirements
-            signedAgreement={project.signed_agreement}
-            approved={project.approved === true}
-            reachedMinFunding={amountRaised >= project.min_funding}
-            projectSlug={project.slug}
-          />
-        )}
-      <div className="flex flex-col gap-4 px-4 pt-5">
-        <ProjectCardHeader
-          round={project.rounds}
-          projectType={project.type}
-          projectTransfer={
-            pendingProjectTransfers?.length === 0
-              ? undefined
-              : project.project_transfers[0]
-          }
-          creator={project.profiles}
-          valuation={isNaN(valuation) ? undefined : valuation}
-          creatorEmail={creatorEmail}
-        />
-        <div>
-          <h2 className="text-3xl font-bold">{project.title}</h2>
-        </div>
-        {project.description && (
-          <Description>
-            <RichContent content={project.description} />
-          </Description>
-        )}
-        {isOwnProject && <EditDescription project={project} />}
-        {project.stage === 'proposal' && (
-          <>
-            <Divider />
-            <ProposalData project={project} raised={amountRaised} />
-          </>
-        )}
-        {(project.stage === 'proposal' ||
-          (project.stage === 'active' && project.type === 'grant')) && (
-          <ProgressBar
-            amountRaised={amountRaised}
-            minFunding={project.min_funding}
-            fundingGoal={project.funding_goal}
-          />
-        )}
-        {profile !== null && project.type === 'cert' && (
-          <PlaceBid
-            project={project}
-            userSpendableFunds={userSpendableFunds}
-            userSellableShares={userSellableShares}
-            userShares={userShares}
-          />
-        )}
-        {profile !== null &&
-          project.type === 'grant' &&
-          pendingProjectTransfers.length === 0 &&
-          (project.stage === 'proposal' || project.stage === 'active') && (
-            <DonateBox
-              project={project}
-              profile={profile}
-              maxDonation={userSpendableFunds}
-            />
-          )}
-        {!user && <SignInButton />}
-        <div id="tabs">
-          <ProjectTabs
-            project={project}
-            user={profile}
-            comments={comments}
-            bids={projectBids}
-            txns={projectTxns}
-            userSpendableFunds={userSpendableFunds}
-            userSellableShares={userSellableShares}
-          />
-        </div>
-        {isAdmin(user) && <CloseBidding project={project} />}
-      </div>
+      <ProjectDisplay
+        project={project}
+        userTxns={userTxns}
+        comments={comments}
+        projectBids={projectBids}
+        projectTxns={projectTxns}
+        creatorEmail={creatorEmail}
+        userProfile={userProfile ?? undefined}
+      />
+      {isAdmin(user) && <CloseBidding project={project} />}
     </>
   )
 }
